+++ src/main/java/region/jidogam/domain/user/controller/UserController.java
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import region.jidogam.common.dto.response.ResponseDto;
 import region.jidogam.common.util.CookieUtil;
-import region.jidogam.domain.jwt.dto.TokenPair;
-import region.jidogam.domain.jwt.dto.TokenResponse;
+import region.jidogam.infrastructure.jwt.dto.TokenPair;
+import region.jidogam.infrastructure.jwt.dto.TokenResponse;
 import region.jidogam.domain.user.dto.UserCreateRequest;
 import region.jidogam.domain.user.service.UserService;
