+++ src/main/java/region/jidogam/domain/user/service/UserService.java
@@ -5,9 +5,9 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import region.jidogam.domain.jwt.JwtProvider;
-import region.jidogam.domain.jwt.RefreshTokenService;
-import region.jidogam.domain.jwt.dto.TokenPair;
+import region.jidogam.infrastructure.jwt.JwtProvider;
+import region.jidogam.infrastructure.jwt.RefreshTokenService;
+import region.jidogam.infrastructure.jwt.dto.TokenPair;
 import region.jidogam.domain.user.dto.UserCreateRequest;
 import region.jidogam.domain.user.entity.User;
 import region.jidogam.domain.user.exception.UserEmailConflictException;
