+++ src/main/java/region/jidogam/infrastructure/jwt/RefreshTokenRepository.java
@@ -1,4 +1,4 @@
-package region.jidogam.domain.jwt;
+package region.jidogam.infrastructure.jwt;
 
 import java.util.Optional;
 import java.util.UUID;
