+++ src/main/java/region/jidogam/infrastructure/jwt/dto/TokenResponse.java
@@ -1,4 +1,4 @@
-package region.jidogam.domain.jwt.dto;
+package region.jidogam.infrastructure.jwt.dto;
 
 public record TokenResponse (
     String accessToken
