+++ src/test/java/region/jidogam/domain/user/service/UserServiceTest.java
@@ -15,10 +15,10 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import region.jidogam.domain.jwt.JwtProvider;
-import region.jidogam.domain.jwt.RefreshToken;
-import region.jidogam.domain.jwt.RefreshTokenService;
-import region.jidogam.domain.jwt.dto.TokenPair;
+import region.jidogam.infrastructure.jwt.JwtProvider;
+import region.jidogam.infrastructure.jwt.RefreshToken;
+import region.jidogam.infrastructure.jwt.RefreshTokenService;
+import region.jidogam.infrastructure.jwt.dto.TokenPair;
 import region.jidogam.domain.user.dto.UserCreateRequest;
 import region.jidogam.domain.user.entity.User;
 import region.jidogam.domain.user.exception.UserEmailConflictException;
